+++ src/js/Editors/index.js
@@ -2,6 +2,7 @@ import m from 'mithril'
 
 import Alert from '../Components/Alert'
 import { Box } from '../Components/Box'
+import Icon from '../Components/Icon'
 import EntryPreview from './EntryPreview'
 import AdvancedOptions from './AdvancedOptions'
 import Proxy from '../Controllers/Proxy'
@@ -111,7 +112,7 @@ const Editor = ({ attrs }) => {
 	return {
 		view: () =>
 			m(Box, {
-				icon: attrs.icon, // '.far.fa-note-sticky',
+				icon: attrs.icon, // 'note',
 				title: postType?.name || attrs.title // 'Note'
 			}, m('form', {
 				onsubmit: post
@@ -140,8 +141,8 @@ const Editor = ({ attrs }) => {
 									selector: 'button.xs',
 									href: '/new/image',
 									disabled: !mediaEndpoint,
-									title: !mediaEndpoint ? 'media-endpoint not found' : ''
-								}, m('i.fas.fa-cloud-arrow-up', { title: 'upload' }))
+									title: !mediaEndpoint ? 'media-endpoint not found' : 'upload'
+								}, m(Icon, { name: 'cloud-arrow-up', label: 'upload' }))
 							]),
 							m('li', m('input', {
 								type: 'text',
@@ -176,17 +177,21 @@ const Editor = ({ attrs }) => {
 									onclick: async (e) => {
 										e && e.preventDefault()
 										state.fetching = true
-										const res = await m
-											.request({
-												method: 'GET',
-												url: '/api/opengraph',
-												params: { url: state[c.type] }
-											})
-										state.name = res.title
+										try {
+											const res = await m
+												.request({
+													method: 'GET',
+													url: '/api/opengraph',
+													params: { url: state[c.type] }
+												})
+											state.name = res.title
+										} catch(err) {
+											Alert.error(err)
+										}
 										state.fetching = false
 									},
-									disabled: state.fetching
-								}, state.fetching ? m('i.fas.fa-spinner.fa-spin', { 'aria-hidden': 'true' }) : m('i.fas.fa-search'))
+									disabled: !state[c.type] || state.fetching
+								}, state.fetching ? m(Icon, { name: 'spinner', className: 'spin' }) : m(Icon, { name: 'magnifying-glass', label: 'get page title' }))
 							])
 						])
 					case 'rsvp':
@@ -216,7 +221,7 @@ const Editor = ({ attrs }) => {
 				m('div.text-center', m('button', {
 					type: 'submit',
 					disabled: state.submitting
-				}, state.submitting ? m('i.fas.fa-spinner.fa-spin', { 'aria-hidden': 'true' }) : 'Post')),
+				}, state.submitting ? m(Icon, { name: 'spinner', clasName: 'spin' }) : 'Post')),
 				m(EntryPreview, { buildPreview: buildEntry })
 			]))
 	}